requirements.txt
torch>=1.9.0
torchvision>=0.10.0
diffusers>=0.12.0
transformers>=4.25.1
accelerate>=0.16.0
numpy>=1.21.0
opencv-python>=4.5.5
pillow>=9.0.0
matplotlib>=3.5.0
pyyaml>=6.0
scipy>=1.7.3
einops>=0.4.1
tqdm>=4.64.0
omegaconf>=2.2.3
gradio>=3.16.0
tensorboard>=2.10.0
safetensors>=0.3.0
librosa>=0.9.2
moviepy>=1.0.3
flask>=2.2.0
scikit-learn>=1.0.2
attrs>=22.1.0
fastapi>=0.89.0
uvicorn>=0.20.0
dearpygui>=1.6.2
xformers>=0.0.16
wandb>=0.15.0